Implement conditions for EC config secret
