+++ pkg/api/v1alpha2/edgeconnect/edgeconnect_types.go
@@ -5,8 +5,6 @@
 package edgeconnect
 
 import (
-	"strings"
-
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/shared/image"
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/shared/proxy"
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/status"
@@ -169,38 +167,4 @@ const (
 
 func init() {
 	v1alpha2.SchemeBuilder.Register(&EdgeConnect{}, &EdgeConnectList{})
-}
-
-func (e *EdgeConnect) HostPatterns() []string {
-	if !e.IsK8SAutomationEnabled() {
-		return e.Spec.HostPatterns
-	}
-
-	var hostPatterns []string
-
-	for _, hostPattern := range e.Spec.HostPatterns {
-		if !strings.EqualFold(hostPattern, e.K8sAutomationHostPattern()) {
-			hostPatterns = append(hostPatterns, hostPattern)
-		}
-	}
-
-	hostPatterns = append(hostPatterns, e.K8sAutomationHostPattern())
-
-	return hostPatterns
-}
-
-type HostMapping struct {
-	From string `json:"from"`
-	To   string `json:"to"`
-}
-
-func (e *EdgeConnect) HostMappings() []HostMapping {
-	hostMappings := make([]HostMapping, 0)
-	hostMappings = append(hostMappings, HostMapping{From: e.K8sAutomationHostPattern(), To: KubernetesDefaultDNS})
-
-	return hostMappings
-}
-
-func (e *EdgeConnect) K8sAutomationHostPattern() string {
-	return e.Name + "." + e.Namespace + "." + e.Status.KubeSystemUID + "." + kubernetesHostnameSuffix
 }
