+++ pkg/api/v1alpha2/edgeconnect/properties.go
@@ -2,6 +2,7 @@ package edgeconnect
 
 import (
 	"fmt"
+	"strings"
 
 	"github.com/Dynatrace/dynatrace-operator/pkg/api"
 	corev1 "k8s.io/api/core/v1"
@@ -50,4 +51,40 @@ func (edgeConnect *EdgeConnect) EmptyPullSecret() corev1.Secret {
 			Namespace: edgeConnect.Namespace,
 		},
 	}
+}
+
+func (ec *EdgeConnect) Conditions() *[]metav1.Condition { return &ec.Status.Conditions }
+
+func (e *EdgeConnect) HostPatterns() []string {
+	if !e.IsK8SAutomationEnabled() {
+		return e.Spec.HostPatterns
+	}
+
+	var hostPatterns []string
+
+	for _, hostPattern := range e.Spec.HostPatterns {
+		if !strings.EqualFold(hostPattern, e.K8sAutomationHostPattern()) {
+			hostPatterns = append(hostPatterns, hostPattern)
+		}
+	}
+
+	hostPatterns = append(hostPatterns, e.K8sAutomationHostPattern())
+
+	return hostPatterns
+}
+
+type HostMapping struct {
+	From string `json:"from"`
+	To   string `json:"to"`
+}
+
+func (e *EdgeConnect) HostMappings() []HostMapping {
+	hostMappings := make([]HostMapping, 0)
+	hostMappings = append(hostMappings, HostMapping{From: e.K8sAutomationHostPattern(), To: KubernetesDefaultDNS})
+
+	return hostMappings
+}
+
+func (e *EdgeConnect) K8sAutomationHostPattern() string {
+	return e.Name + "." + e.Namespace + "." + e.Status.KubeSystemUID + "." + kubernetesHostnameSuffix
 }
