+++ pkg/controllers/edgeconnect/conditions.go
@@ -1,0 +1,4 @@
+package edgeconnect
+
+// SecretConfigConditionType identifies the secret config condition.
+const SecretConfigConditionType = "SecretConfigConditionType"
